refactor recorder into separate files
